Fixing infra issues, exporting google branch changes.
FUTURE_COPYBARA_INTEGRATE_REVIEW=https://github.com/google/iree/pull/4006 from rsuderman:llvm-dependent-submodule-update 876dfddd688bfcb201308207a605050a7bda65a2
PiperOrigin-RevId: 344872931

+++ iree/hal/host/host_executable.h
@@ -41,7 +41,7 @@ class HostExecutable : public Executable {
   // Grid parameters shared for all tiles within a dispatch.
   struct DispatchParams {
     // Entry point within the executable.
-    int32_t entry_point = 0;
+    size_t entry_point = 0;
 
     // Total workgroup XYZ count for the grid.
     std::array<uint32_t, 3> workgroup_count;
